+++ crates/zng-task/src/channel.rs
@@ -36,7 +36,7 @@ mod ipc;
 pub use ipc::{IpcReceiver, IpcSender, IpcValue, NamedIpcReceiver, NamedIpcSender, ipc_unbounded};
 
 mod ipc_bytes;
-pub use ipc_bytes::{IpcBytes, IpcBytesMut, IpcBytesWriter, IpcBytesWriterBlocking, WeakIpcBytes};
+pub use ipc_bytes::{IpcBytes, IpcBytesCast, IpcBytesMut, IpcBytesMutCast, IpcBytesWriter, IpcBytesWriterBlocking, WeakIpcBytes};
 
 #[cfg(ipc)]
 pub use ipc_bytes::{is_ipc_serialization, with_ipc_serialization};